move function in PoximityChatRoom
