+++ play/src/front/Chat/Connection/Proximity/ProximityChatRoom.ts
@@ -19,12 +19,13 @@ import { iframeListener } from "../../../Api/IframeListener";
 import { SpaceInterface } from "../../../Space/SpaceInterface";
 import { SpaceRegistryInterface } from "../../../Space/SpaceRegistry/SpaceRegistryInterface";
 import { chatVisibilityStore } from "../../../Stores/ChatStore";
-import { selectedRoom } from "../../Stores/ChatStore";
+import { isAChatRoomIsVisible, navChat, selectedRoom } from "../../Stores/ChatStore";
 import { SpaceFilterInterface, SpaceUserExtended } from "../../../Space/SpaceFilter/SpaceFilter";
 import { mapExtendedSpaceUserToChatUser } from "../../UserProvider/ChatUserMapper";
 import { SimplePeer } from "../../../WebRtc/SimplePeer";
 import { bindMuteEventsToSpace } from "../../../Space/Utils/BindMuteEvents";
 import { gameManager } from "../../../Phaser/Game/GameManager";
+import { availabilityStatusStore, mediaStreamConstraintsStore } from "../../../Stores/MediaStore";
 
 export class ProximityChatMessage implements ChatMessage {
     isQuotedMessage = undefined;
@@ -383,6 +384,20 @@ export class ProximityChatRoom implements ChatRoom {
         });
 
         this.simplePeer.setSpaceFilter(this._spaceWatcher);
+
+        const mediaStreamConstraints = get(mediaStreamConstraintsStore);
+        const actualStatus = get(availabilityStatusStore);
+        if (!isAChatRoomIsVisible()) {
+            selectedRoom.set(this);
+            navChat.set("chat");
+            if (
+                !mediaStreamConstraints.audio &&
+                !mediaStreamConstraints.video &&
+                (actualStatus === AvailabilityStatus.ONLINE || actualStatus === AvailabilityStatus.AWAY)
+            ) {
+                chatVisibilityStore.set(true);
+            }
+        }
     }
 
     public leaveSpace(spaceName: string): void {
